+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.7.1-common-dev
 
+- Fix conversion warnings in iOS version
 - Fix to Windows "Universal" version to support big integers
 - Implement database close and delete operations for Windows "Universal"
 - Fix readTransaction to skip BEGIN/COMMIT/ROLLBACK